scripts/pyproject.toml
[tool.poetry]
name = "fortytwo-fetcher"
version = "1.0.0"
description = ""
authors = ["lucas-ht <46166712+lucas-ht@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
idna = "3.7"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
requests = "2.31.0"
six = "1.16.0"
urllib3 = "2.2.1"
numpy = "^2.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "0.9.0"


[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "B",
    "C",
    "I",
    "TCH",
    "ANN",
]
ignore = [
    "ANN401"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
